D FLip Flop sync to clk
The problem with sync is reset signal. Lets say reset is active for less than 1 clock cycle. In that case it is possible that when reset signal happend we are not sure if our clock will catch it, Hence we see two signals with such versions. One on positive edge of the clock, and the other on negative edge of the clock. The OR result both is latched on for actual reset, which arrives a clock late!!!! So an extra clock cycle to ensure the signal is indeed caught.